refactor(mcp): P6-B4 second cut — extract mcp_compat.{hpp,cpp}
Phase 6 B4 (MCP typed-registry split), second cut. The project
SDK-compatibility layer — which resolves a project's pinned SDK and
gates tools whose min-SDK floor the project hasn't reached —
relocated into its own TU.

mcp_compat.hpp — public surface (5 fns):
  * min_sdk_for_tool / resolve_project_sdk_version
  * compare_semver / compat_error_payload / handle_compat

mcp_compat.cpp — public defs + file-local statics (pulp.toml /
CMakeLists SDK-version scanners, semver triple parser, ToolMinSdk
table). Includes mcp_json.hpp for the JSON-RPC framing.

pulp_mcp.cpp pulls the 5 public fns into file scope via
using-declarations so existing call sites are unchanged.

pulp_mcp.cpp: 844 → 597 (-247; 951 → 597 across both B4 cuts so far).
Verified: pulp-mcp builds clean; pulp_compat tool round-trips a valid
structured payload (project_sdk + version fields populated).
