+++ tools/mcp/CMakeLists.txt
@@ -8,7 +8,9 @@ configure_file(
     "${CMAKE_CURRENT_BINARY_DIR}/pulp_mcp_version.h"
     @ONLY)
 
-add_executable(pulp-mcp pulp_mcp.cpp)
+add_executable(pulp-mcp
+    pulp_mcp.cpp
+    mcp_compat.cpp)
 set_target_properties(pulp-mcp PROPERTIES OUTPUT_NAME "pulp-mcp")
 target_compile_features(pulp-mcp PRIVATE cxx_std_20)
 target_include_directories(pulp-mcp PRIVATE "${CMAKE_CURRENT_BINARY_DIR}")
